+++ wpull/network.py
@@ -11,7 +11,7 @@ import tornado.gen
 import wpull.async
 from wpull.cache import FIFOCache
 from wpull.errors import NetworkError, DNSNotFound
-import wpull.util
+import wpull.string
 
 
 _logger = logging.getLogger(__name__)
@@ -91,7 +91,7 @@ class Resolver(object):
         if not addresses:
             raise DNSNotFound(
                 "DNS resolution for '{0}' did not return any results."\
-                .format(wpull.util.coerce_str_to_ascii(host))
+                .format(wpull.string.coerce_str_to_ascii(host))
             )
 
         _logger.debug('Resolved addresses: {0}.'.format(addresses))