+++ doc/changelog.rst
@@ -13,6 +13,7 @@ What's New
   * Moves ``OrderedDefaultDict`` from ``util`` to ``collections``.
   * Moves ``DeflateDecompressor``, ``gzip_decompress`` from ``util`` to ``decompression``.
   * Moves ``sleep``, ``TimedOut``, ``wait_future``, ``AdjustableSemaphore`` from ``util`` to ``async``.
+  * Moves ``to_bytes``, ``to_str``, ``normalize_codec_name``, ``detect_encoding``, ``try_decoding``, ``format_size``, ``printable_bytes``, ``coerce_str_to_ascii`` from ``util`` to ``string``.
 
 
 0.32 (2014-04-17)